pyproject.toml
[tool.poetry]
name = "log_analysis"
version = "0.1.0"
description = ""
authors = ["Gavin Lin <gavinjwl@amazon.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
redshift-connector = "^2.0.907"
boto3 = "^1.0.0"
aws-cdk-lib = "^2.0.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pytest = "^7.1.2"
pylint = "^2.13.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
